JsonRead implemented but needs some testing.

+++ app/src/main/java/com/lovelace_scd/timed/model/Medication.kt
@@ -1,8 +1,8 @@
 package com.lovelace_scd.timed.model
 
 class Medication(var name: String, var rxFullSize: Double, var amountRemaining: Double,
-                      var numRefillsRemaining: Int, var isRefillable: Boolean, var dosesPerTimePeriod: Int,
-                      var daysPerTimePeriod: Int, var doseSize: Double, var takeWithFood: Boolean, var doseUnit: String ) {
+                      var numRefillsRemaining: Int, var isRefillable: Boolean, var dosesPerTimePeriod: Long,
+                      var daysPerTimePeriod: Long, var doseSize: Double, var takeWithFood: Boolean, var doseUnit: String ) {
 
     private var footTaken : Boolean = false
